Remove semivariogram; move affine ops to package ops
