+++ vector/src/main/scala/geotrellis/vector/op/affine/README.md
@@ -1,4 +1,4 @@
-#geotrellis.vector.affine
+#geotrellis.vector.op.affine
 
 The contents of this package define affine transformations over the geometric entities of `geotrellis.vector`. [Affine transformations](http://mathworld.wolfram.com/AffineTransformation.html) preserve the proportions of lines (both those that constitute the geometry of interest and any lines which *could* be drawn between them).
